+++ ShipSandbox TODO.txt
@@ -120,31 +120,46 @@ NOTE: fps drop is wholly due to additional vertex shader outputs; it has
 			+ ...do tension strain check...
 			+ ...electrical and water...
 		- Re-introduce:
-			= Masses				
-				- Gamma - from original
+			+ Masses				
+			+ Gamma - from original (NOT from original, doesn't work)
+			- Better grab force:
+				- Have optional<float force, vec2f position>; at each iteration, if not none:
+					- Do force spreading with force/NumIterations
+				+ Also, all non-static GameParameters initialized at cctor in cpp				
 				- Check ropes afterwards
+			- Put spring damping factor in GameParameters as well
 			- Collision with sea floor and (original) "bounce"
-			- Adjust better grab force based off new static game parameter: NumIterations
-				- Also, all non-static GameParameters initialized at cctor in cpp
-				- Should spread grab force over iterations? Otherwise it's pulses
-					- Could have optional<float force, vec2f position>; at each iteration, if not none:
-						- Do force spreading with force/NumIterations
+				- v += adjusted Pos - pre-adjusted Pos
+		- Test:
+			- Self-adjustment of kDamp based off paper's formula (H and L bounds)
+				- Debug to see if bounds ever exceeded
 		- Perf:
 			- Get baseline (fps and trace)
-			- Do Point.PreparedMass
+			- Do Point.MassFactor
 				- const calc'd at cctor via dt passed to cctor, which was passed by Ship::Create via gameParameters
 			- Get baseline (fps)
-			- Do Spring.PreparedXYZ
+			- Do Spring.XYZFactor/s
 			- Get baseline (fps and trace)
 			- To enable SSE2, should use 4 floats (or 2 doubles)
 					- Test 4 floats:
 						- Test 1: Integrate: 
 							- 1a: Do two points at a time, with twin instructions
+								- For int loop step 2 points, no remainder - see below
 							- 1b: pack Pos, Vel, Force from 2 points into vec4f
-							- If ok: see if worth revisiting Pos, Vel, and Force into own vectors pointed to by Point
-					- If we go for doubles:
+								- Same for loop
+						- If ok: see if worth revisiting Pos, Vel, and Force into own vectors pointed to by Point
+							- Ship has 5 vectors of unions (float[4], {vec2f cPoint1, vec2f cPoint2}):
+							  Positions, Velocities, Forces, and Masses (or MassFactors), all
+							  allocated and calc'd at Initialize()
+								- Masses have another union, with 4 separate floats, and are const
+								- Rounded to 2 (static constexpr = 16 / 2*sizeof(float) (16 = 128bit SSE, 2 = coordinates in 2D) with 0's
+								- Masses are rounded to 4 (static constexpr = 16 / 1*sizeof(float) (16 = 128bit SSE, 1 = #masses) with 0's
+							- Point has vec2f* Position, vec2f* Velocity, vec2f* Force 
+								- But own float Mass, it's const anyway, and better for storage and CPU to store
+							- Integrate builds own vector of TimeSteps(/Factors)
+					- If we want to test for doubles:
 						- SimulationStep: templated constant (double and float)
-						- If ok, should expand to Mass, Buoyancy? 
+						
 		- If it works (!!!):
 			- Set springs' K in material
 				- Fix Titanic masts
